requirements.txt
streamlit
transformers
torch
openai-whisper
ffmpeg-python
